Move fixtures into fixture folder
