+++ plugins/woocommerce/tests/e2e-pw/fixtures/block-editor-fixtures.js
