+++ plugins/woocommerce/tests/e2e-pw/fixtures/fixtures.js
