+++ plugins/woocommerce/tests/e2e-pw/tests/merchant/product-create-simple.spec.js
@@ -1,4 +1,4 @@
-const { test: baseTest, expect } = require( '../../fixtures' );
+const { test: baseTest, expect } = require( '../../fixtures/fixtures' );
 
 baseTest.describe( 'Products > Add Simple Product', () => {
 	const productData = {
