+++ plugins/woocommerce/tests/e2e-pw/tests/merchant/product-delete.spec.js
@@ -1,4 +1,4 @@
-const { test: baseTest, expect } = require( '../../fixtures' );
+const { test: baseTest, expect } = require( '../../fixtures/fixtures' );
 
 baseTest.describe( 'Products > Delete Product', () => {
 	const test = baseTest.extend( {
