+++ plugins/woocommerce/tests/e2e-pw/tests/merchant/product-edit.spec.js
@@ -1,4 +1,4 @@
-const { test: baseTest, expect } = require( '../../fixtures' );
+const { test: baseTest, expect } = require( '../../fixtures/fixtures' );
 
 baseTest.describe( 'Products > Edit Product', () => {
 	const test = baseTest.extend( {
