+++ plugins/woocommerce/tests/e2e-pw/tests/merchant/product-linked-producs.spec.js
@@ -1,4 +1,4 @@
-const { test: baseTest, expect } = require( '../../fixtures' );
+const { test: baseTest, expect } = require( '../../fixtures/fixtures' );
 
 baseTest.describe( 'Products > Related products', () => {
 	const test = baseTest.extend( {
