+++ plugins/woocommerce/tests/e2e-pw/tests/merchant/product-reviews.spec.js
@@ -1,4 +1,4 @@
-const { test: baseTest, expect } = require( '../../fixtures' );
+const { test: baseTest, expect } = require( '../../fixtures/fixtures' );
 
 baseTest.describe( 'Product Reviews > Edit Product Review', () => {
 	const test = baseTest.extend( {
